fix: 앱 시작 시 네비게이션 경쟁 상태 오류 해결
